+++ docs/src/pages/index.tsx
@@ -1,0 +1,5 @@
+import { Redirect } from "react-router-dom";
+
+export default function Home() {
+  return <Redirect to="/pocket-id/introduction" />;
+}